requirements.txt
black==21.4b2
coverage==5.5
flake8==3.9.1
flake8-docstrings==1.6.0
Flask==1.1.2
pandas==1.2.4
Pillow==8.2.0
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.0.0
pytest==6.2.3
pytest-cov==2.11.1
python-docx==0.8.10
requests==2.25.1
requests-mock==1.9.2
toml==0.10.2
setuptools-scm==6.0.1
wheel==0.36.2
docx==0.2.4
Werkzeug==2.0.1